Moved alternative response processing to file-uploader/*
